micro refact
