+++ conf/httpdispacher.toml
@@ -5,7 +5,6 @@ domains = ["api.weibo.cn.","weibo.cn.","taobao.com.","www.baidu.com.","www.taoba
 #bool
 mysql_enable = false
 #string
-ipdb_path = "/Users/chunsheng/Dropbox/Work/Sina/08.Projects/16.httpDispacher/data/ip.db.db"
 server_log = "./httpdispacher.server.log"
 query_log = "./httpdispacher.query.log"
 querylog_format = "%{time:2006-01-02T15:04:05} %{shortfile}|%{shortfunc} %{level:.4s} %{id:03x}%{message}"
