+++ src/query/wrapper_test.go
@@ -1,4 +1,4 @@
-package server
+package query
 
 import (
 	"testing"
